Add configurable pause
Problem: When using rspec-eventually it evaluates matcher with default sleep time of 100ms. When using it with API or other HTTP calls it seems excessive to retry every 100ms.

Solution: add ability to configure time to pause between retries as a Rspec::Eventually.pause or pause_for(X) seconds

Most of the tests that I could think of are included.